sdk: evict unchecked_div_by_const
