+++ sdk/program/Cargo.toml
@@ -81,6 +81,7 @@ solana-stable-layout = { workspace = true }
 solana-system-interface = { workspace = true, features = ["bincode"] }
 solana-sysvar = { workspace = true, features = ["bincode", "bytemuck"] }
 solana-sysvar-id = { workspace = true }
+solana-unchecked-div-by-const = { workspace = true }
 thiserror = { workspace = true }
 
 # This is currently needed to build on-chain programs reliably.
