+++ sdk/unchecked-div-by-const/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "solana-unchecked-div-by-const"
+description = "Compile-time-safe division."
+documentation = "https://docs.rs/solana-rent"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[lints]
+workspace = true
